use already existing value_length() function to get usize

+++ src/chunk/chunk_init.rs
@@ -132,7 +132,7 @@ impl Chunk for ChunkInit {
 
         if !(header.typ == CT_INIT || header.typ == CT_INIT_ACK) {
             return Err(Error::ErrChunkTypeNotTypeInit);
-        } else if (header.value_length as usize) < INIT_CHUNK_MIN_LENGTH {
+        } else if header.value_length() < INIT_CHUNK_MIN_LENGTH {
             // validity of value_length is checked in ChunkHeader::unmarshal
             return Err(Error::ErrChunkValueNotLongEnough);
         }
